requirements.txt
absl-py==2.3.1
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.4.0
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.14.2
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
curl_cffi==0.13.0
dataclasses-json==0.6.7
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.104.1
flatbuffers==25.9.23
frozendict==2.4.7
frozenlist==1.8.0
gast==0.6.0
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.6
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-generativeai==0.7.2
google-pasta==0.2.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
h5py==3.15.1
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
keras==2.15.0
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-google-genai==1.0.10
langchain-text-splitters==0.2.4
langsmith==0.1.147
libclang==18.1.1
lxml==6.0.2
Mako==1.3.10
Markdown==3.10
MarkupSafe==3.0.3
marshmallow==3.26.1
ml-dtypes==0.2.0
multidict==6.7.0
multitasking==0.0.12
mypy_extensions==1.1.0
numpy==1.26.3
oauthlib==3.3.1
opencv-python==4.9.0.80
opt_einsum==3.4.0
orjson==3.11.4
packaging==23.2
pandas==2.3.3
passlib==1.7.4
peewee==3.18.3
Pillow==10.1.0
platformdirs==4.5.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.5.0
pydantic_core==2.14.1
pyparsing==3.2.5
pypdf==6.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-slugify==8.0.1
pytz==2025.2
PyYAML==6.0.3
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.23
starlette==0.27.0
tenacity==8.5.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.2.0
text-unidecode==1.3
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.24.0
watchfiles==1.1.1
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.14.2
yarl==1.22.0
yfinance==0.2.66
